+++ stepper/door.h
@@ -1,0 +1,10 @@
+//
+// Created by dfkdream on 22. 12. 9.
+//
+
+#ifndef SMART_DOORLOCK_DOOR_H
+#define SMART_DOORLOCK_DOOR_H
+
+void door(int);
+
+#endif //SMART_DOORLOCK_DOOR_H
